fix: Switch to a version of Mocktopus that doesn't generate Clippy errors
fix: Correct formatting

test: Add a container to simplify running tests

ci: Use LLVM tools to collect code coverage

ci: Switch back to using cargo test for test runs

ci: Don't run coverage collection on non-coverage tests

ci: Only capture coverage reports for git-tool methods

tweak: Allow us to collect coverage on any build machine

ci: Update coverage collection logic

ci: Use pwsh instead of powershell

ci: Use the correct path to the coverage asset

ci: Only install rustfilt if it is required for coverage

fix: Cleanup a remaining code warning

ci: Use the correct rust cache for each toolchain and target

refactor: Cleanup the last remaining clippy warnings:

+++ src/main.rs
@@ -119,7 +119,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
                 }
 
                 tracing::Span::current()
-                    .record("status_code", &(1 as u32))
+                    .record("status_code", &1_u32)
                     .record("otel.status", &field::debug(StatusCode::Error))
                     .record("error", &field::display(&err));
